refactor(info): extract ReadmePane body from InfoPane
Move README fetch/render logic verbatim into ReadmePane (body-only, no
Pane chrome). InfoPane delegates to it — rendered DOM is identical.
